Remove eval_util module from public namespace
Deletes some redundant functionality, moves the rest util
tests/tsutil.py for now.
